Moved reference checking code into individual commands

+++ xarta/commands/delete.py
@@ -3,11 +3,12 @@
 
 import os
 
-from .base import Base
+from .base import BaseCommand
 from ..database import PaperDatabase
+from ..utils import process_ref, is_valid_ref, XartaError
 
 
-class Delete(Base):
+class Delete(BaseCommand):
     """ Remove an arXiv paper and its metadata from the database. """
 
     def run(self):
@@ -15,4 +16,8 @@ class Delete(Base):
         ref = options["<ref>"]
 
         with PaperDatabase() as paper_database:
-            paper_database.delete_paper(ref)
+            ref = process_ref(ref)
+            if is_valid_ref(ref):
+                paper_database.delete_paper(ref)
+            else:
+                raise XartaError("Not a valid arXiv reference or alias: " + ref)
